server/requirements.txt
absl-py==2.0.0
aiohttp==3.7.4.post0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
exceptiongroup==1.1.3
executing==2.0.1
fake-useragent==1.3.0
feedparser==6.0.10
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.3.8
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mindsdb-sdk==2.0.0
mindsdb-sql==0.7.3
ml-dtypes==0.2.0
multidict==6.0.4
multitasking==0.0.11
mysqlclient==2.1.1
nest-asyncio==1.5.8
newsapi-python==0.2.7
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
parse==1.19.1
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pickle5==0.0.11
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==8.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
pyrate-limiter==3.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
sec-edgar-api==1.1.0
sgmllib3k==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.7
vaderSentiment==3.3.2
virtualenv==20.24.6
w3lib==2.1.2
wcwidth==0.2.9
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
wrapt==1.14.1
yahoo-fin==0.8.9.1
yarl==1.9.2
yfinance==0.2.31
zipp==3.17.0